feat(gen): add x-oapi-codegen-extra-tags support to the parameter
