+++ gen/_template/parameters.tmpl
@@ -19,7 +19,8 @@
 type {{ $op.Name }}Params struct {
 {{- range $p := $op.Params }}
 	{{- template "godoc" $p.GoDoc }}
-	{{ $p.Name }} {{ $p.Type.Go }}
+	{{ $pTags := $p.Tag.GetTags }}
+	{{ $p.Name }} {{ $p.Type.Go }} {{ if $pTags }}{{ backquote $pTags }}{{ end }}
 {{- end }}
 }
